fix: compile(Linux)

+++ math/ft_round.c
@@ -6,14 +6,14 @@
 /*   By: stanaka2 <stanaka2@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/30 18:43:47 by stanaka2          #+#    #+#             */
-/*   Updated: 2026/06/01 20:40:13 by stanaka2         ###   ########.fr       */
+/*   Updated: 2026/06/04 14:30:19 by stanaka2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 double	ft_round(double x)
 {
 	if (x >= 0)
-		return ((long long)(x + 0.5));
+		return ((double)(long long)(x + 0.5));
 	else
-		return ((long long)(x - 0.5));
+		return ((double)(long long)(x - 0.5));
 }
